Lesson 6

+++ authapp/models.py
@@ -5,7 +5,7 @@ from mainapp.models import NULLABLE
 
 class User(AbstractUser):
     email = models.EmailField(verbose_name='Email', blank=True, unique=True)
-    age = models.PositiveIntegerField(verbose_name='возраст', **NULLABLE)
+    age = models.PositiveSmallIntegerField(verbose_name='возраст', **NULLABLE)
     avatar = models.ImageField(upload_to='users', **NULLABLE)
 
     class Meta:
